protocol tester cleanup
